fix dex-auth configmap's config.yaml formatting
#31 outlines issues where quotes from the JSON dump of config settings into config-map.yaml are being taken literally, causing login errors.  This commit changes to using yaml (no quotes needed) and modifies the config-map.yaml template to preserve indenting of the pasted yaml.  Without changes to config-map.yaml, the newlines in the yaml inserted into the template break the output, resulting in a configmap that has only a single line of config in it (and as a result, dex does not start).

Code after this commit is NOT fully working.  When dex+oidc-auth+istio are related, browsing to the ingress IP results in a dex page stating: "Bad Request.  Invalid client_id ("authservice-oidc").".  It is unclear what has gone wrong with this setup.  At one point it appeared to work, but after rebuilding the environment and trying again it did not.