build: install bcrypt from binary and pin its version <4.1
Recently, this package's dependencies introduced a dependency on rustc, a toolchain
that is not installed in the charm by default, which can cause trouble when installing
(building and installing) bcrypt from source during the charmcraft pack.
Also pinning bcrypt to <4.1 to keep some sort of control over the version used.