build: update requirements*.in and their corresponding .txt files to match standards (#144)
* build: update requirements*.in and their corresponding .txt files to match standards

Using the same package versions for this repository as we do in others guarantee a certain
standard across all of them. This commit removes explicit package pins and upgrades some
package versions accordingly.

* build: install rustc and cargo as build packages

Some of this charm dependencies need rustc and cargo to build and install from source,
poviding this toolchain and library ensures the right tools are installed in the
environment, preventing issues when pip installing charm dependencies.